Refactor AppInput to wrapper dumb component
